requirements.txt
PyQt6>=6.4.0
sounddevice>=0.4.6
soundfile>=0.12.0
numpy>=1.24.0
openai-whisper>=20231117